ESP12E_Whether_Station
This code is for an ESP8266 microcontroller that collects data from DHT22 and DallasTemperature sensors (for temperature, humidity, and water temperature), and sends the data to a remote server via HTTP POST. It includes Wi-Fi connection, NTP time synchronization, and deep sleep to conserve power. The collected data is formatted as JSON and includes timestamps from the NTP server. The code repeats this process at intervals and provides serial output for debugging.